refactor(carRacing) : Repackage follow by MVC and remove secondary constructor for testing
