+++ src/main/kotlin/carracing/view/InputView.kt
@@ -1,6 +1,8 @@
-package carracing
+package carracing.view
 
-object CarRacingView {
+import carracing.CarRacingException
+
+object InputView {
 
     fun getCarNames(): String {
         println("경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분).")
@@ -17,7 +19,7 @@ object CarRacingView {
         if (inputCount.isBlank()) {
             throw NullPointerException("숫자로 입력해주세요.")
         }
-        var count: Int = 0
+        var count: Int
         try {
             count = inputCount.toInt()
         } catch (e: NumberFormatException) {
@@ -27,21 +29,5 @@ object CarRacingView {
             throw CarRacingException("자연수를 입력해주세요.")
         }
         return count
-    }
-
-    fun showCarMovement(cars: MutableList<Car>) {
-        cars.forEach {
-            print(it.name + " : " + "-".repeat(it.position) + "\n")
-        }
-    }
-
-    fun showWinners(winners: List<Car>) {
-        val result = StringBuilder()
-        winners.forEach {
-            result.append(it.name).append(", ")
-        }
-        result.delete(result.length - 2, result.length)
-
-        print(result.toString() + "가 최종 우승했습니다.")
     }
 }
